update get_job handler
