+++ api/api.py
@@ -1,11 +1,5 @@
-from flask import Flask
-from flask import request
-from spider.args import get_provider, get_provider_handler
-from scrapy.crawler import CrawlerProcess
-from scrapy.signalmanager import dispatcher
-from scrapy import signals
-from scrapy.utils.project import get_project_settings
-from spider.spiders import SpiderRequest
+from flask import Flask, request
+from spider.spider.crawl import crawl_job
 
 app = Flask(__name__)
 
@@ -23,20 +17,8 @@ def get_job():
         platforms = platforms.split(",")
     if skills:
         skills = skills.split(",")
+    results = crawl_job(
+        job=jobTitle, location=location, skills=skills, platforms=platforms, page=page
+    )
 
-    results = []  # scraped items
-
-    def store_crawler_item(item):
-        results.append(item)
-
-    # store new scraped item to results
-    dispatcher.connect(store_crawler_item, signal=signals.item_scraped)
-    settings = get_project_settings()
-    process = CrawlerProcess(settings=settings)
-    for platform in platforms:
-        spider = get_provider_handler(get_provider(platform))
-        request = SpiderRequest(
-            role=jobTitle, location=location, skills=skills, page=page
-        )
-        process.crawl(spider, request)
-    process.start()
+    return {"data": results}
